requirements.txt
Flask==0.12
flask-restful==0.3.5
Flask-SQLAlchemy==2.4.3
flask-marshmallow==0.14.0
marshmallow-sqlalchemy==0.24.0
nltk==3.5
numpy==1.19.4
PyMySQL==0.10.1
cryptography